Feature/job submission flow

+++ src/main/java/oosd/sait/ipdbuddyapi/entities/BillableItemSubmission.java
@@ -62,7 +62,7 @@ public class BillableItemSubmission {
         return totalPrice;
     }
 
-    public void setTotalPrice(double totalPrice) {
+    public void setTotalPrice(BigDecimal totalPrice) {
         this.totalPrice = totalPrice;
     }
